+++ Assets/Scripts/ClickOnCNC.cs
