+++ Assets/examples/spindleTest.cnc
@@ -1,4 +1,4 @@
-;5/3/2020 8:54:28 PM
+;5/4/2020 3:54:32 PM
 M3 S0
 M3 S1
 M3 S2
